+++ inmanta_plugins/files/systemd_unit.py
@@ -40,8 +40,6 @@ class SystemdUnitFileResource(inmanta_plugins.files.base.BaseFileResource):
 class SystemdUnitFileHandler(
     inmanta_plugins.files.base.BaseFileHandler[SystemdUnitFileResource]
 ):
-    _io: inmanta.agent.io.local.LocalIO
-
     def read_resource(
         self,
         ctx: inmanta.agent.handler.HandlerContext,
@@ -50,7 +48,7 @@ class SystemdUnitFileHandler(
         super().read_resource(ctx, resource)
 
         # Load the content of the existing file
-        resource.content = self._io.read_binary(resource.path).decode()
+        resource.content = self.proxy.read_binary(resource.path).decode()
         ctx.debug("Reading existing file", content=resource.content)
 
     def create_resource(
@@ -58,7 +56,7 @@ class SystemdUnitFileHandler(
         ctx: inmanta.agent.handler.HandlerContext,
         resource: SystemdUnitFileResource,
     ) -> None:
-        self._io.put(resource.path, resource.content.encode())
+        self.proxy.put(resource.path, resource.content.encode())
         super().create_resource(ctx, resource)
 
     def update_resource(
@@ -68,6 +66,6 @@ class SystemdUnitFileHandler(
         resource: SystemdUnitFileResource,
     ) -> None:
         if "content" in changes:
-            self._io.put(resource.path, resource.content.encode())
+            self.proxy.put(resource.path, resource.content.encode())
 
         super().update_resource(ctx, changes, resource)