+++ tests/test_symlink.py
@@ -32,6 +32,7 @@ def test_model(
     user = os.getlogin()
     group = grp.getgrgid(os.getgid()).gr_name
     model = f"""
+        import mitogen
         import files
         import files::host
 
@@ -40,6 +41,7 @@ def test_model(
         host = std::Host(
             name="localhost",
             os=std::linux,
+            via=mitogen::Local(),
         )
 
         files::Symlink(